Part 5. 게시물 <=> 유저 관계 설정 및 게시물 API 수정

+++ src/main/java/com/fastcampus/board/service/PostService.java
@@ -1,7 +1,9 @@
 package com.fastcampus.board.service;
 
 import com.fastcampus.board.exception.post.PostNotFoundException;
+import com.fastcampus.board.exception.user.UserNotAllowedException;
 import com.fastcampus.board.model.entity.PostEntity;
+import com.fastcampus.board.model.entity.UserEntity;
 import com.fastcampus.board.model.post.Post;
 import com.fastcampus.board.model.post.PostPatchRequestBody;
 import com.fastcampus.board.model.post.PostPostRequestBody;
@@ -26,24 +28,34 @@ public class PostService {
     return Post.from(postEntity);
   }
 
-  public Post createPost(PostPostRequestBody postPostRequestBody) {
-    var postEntity = new PostEntity();
-    postEntity.setBody(postPostRequestBody.body());
-    var savedPostEntity = postEntityRepository.save(postEntity);
+  public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+    var savedPostEntity =
+        postEntityRepository.save(PostEntity.of(postPostRequestBody.body(), currentUser));
     return Post.from(savedPostEntity);
   }
 
-  public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+  public Post updatePost(
+      Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
     var postEntity =
         postEntityRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
+
+    if (!postEntity.getUser().equals(currentUser)) {
+      throw new UserNotAllowedException();
+    }
+
     postEntity.setBody(postPatchRequestBody.body());
     var updatedEntity = postEntityRepository.save(postEntity);
     return Post.from(updatedEntity);
   }
 
-  public void deletePost(Long postId) {
+  public void deletePost(Long postId, UserEntity currentUser) {
     var postEntity =
         postEntityRepository.findById(postId).orElseThrow(() -> new PostNotFoundException(postId));
+
+    if (!postEntity.getUser().equals(currentUser)) {
+      throw new UserNotAllowedException();
+    }
+
     postEntityRepository.delete(postEntity);
   }
 }
